Move GitHubAdapter to separate class
